% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_estimators.R
\name{compute_cov_matrix}
\alias{compute_cov_matrix}
\title{Compute Covariance Matrix}
\usage{
compute_cov_matrix(resid, zero_mean)
}
\arguments{
\item{resid}{A numeric matrix of residuals with row obs and col variables.}

\item{zero_mean}{Logical; if TRUE, residuals are assumed to have zero mean.}
}
\description{
Computes the covariance matrix. When residuals are assumed to have
zero mean, use cross-product divided by T; otherwise, cov().
}
