% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_estimators.R
\name{compute_var_cor_matrix}
\alias{compute_var_cor_matrix}
\title{Compute Variance of Sample Correlation}
\usage{
compute_var_cor_matrix(resid, covm, zero_mean)
}
\arguments{
\item{resid}{A numeric matrix of residuals.}

\item{covm}{A covariance matrix, computed by compute_cov_matrix().}

\item{zero_mean}{Logical; if TRUE, residuals are assumed to have zero mean.}
}
\description{
Computes the variance of sample correlation from residuals, which will be
standardised by given covariance matrix. The variance of sample correlation
formula differs depending on zero-mean assumption on residuals.
}
