% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{convert_cor2cov}
\alias{convert_cor2cov}
\title{Convert correlation to covariance matrix}
\usage{
convert_cor2cov(cor, stdevs = runif(nrow(cor), sqrt(2), sqrt(6)))
}
\arguments{
\item{cor}{A square correlation matrix.}

\item{stdevs}{A numeric vector of standard deviations for each variable.}
}
\value{
A covariance matrix.
}
\description{
This function converts a correlation matrix to a covariance matrix
using specified standard deviations.
}
