% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{flip_signs_mat}
\alias{flip_signs_mat}
\title{Flip totally positive matrix to mixed-sign matrix}
\usage{
flip_signs_mat(mat, flip_prob = 0.5, ensure_PD = TRUE)
}
\description{
Randomly flips signs of the elements with probability \code{flip_prob}.
}
