% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_cor}
\alias{generate_cor}
\title{Generate Correlation Structure using Hardin et al. (2013) algorithm 1}
\usage{
generate_cor(
  groups = c(2, 3, 4),
  rho = runif(length(groups), 0.2, 0.7),
  delta = min(rho) * 0.8,
  epsilon = (1 - max(rho)) * 0.8,
  eidim = 2,
  random_seed = NULL
)
}
\description{
Start with constant correlations within each group (rhos) and between
each group (delta). Then add random noise to all elements of the matrix, by
scaling epsilon by outer cross products of random unit vectors.
}
\references{
Hardin, J., Garcia, S. R., & Golan, D. (2013). A method for generating realistic correlation matrices. The Annals of Applied Statistics, 7(3), 1733–1762. https://www.jstor.org/stable/23566492
}
