% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{make_PD_cpp}
\alias{make_PD_cpp}
\title{Enforce positive definiteness on a matrix in C++}
\usage{
make_PD_cpp(W, tol = 1e-08)
}
\arguments{
\item{W}{The input matrix to be made positive definite.}

\item{threshold}{A small value to replace negative eigenvalues (1e-8 default).}
}
\value{
A positive definite matrix.
}
\description{
This function performing an eigen decomposition on a symmetric matrix,
thresholding the eigenvalues, and reconstructing the matrix.
}
