% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{novelist_cov_grid_cpp}
\alias{novelist_cov_grid_cpp}
\title{Outsource the two for loops in novelist_cv into C++}
\usage{
novelist_cov_grid_cpp(resid, deltas, window_size, zero_mean = TRUE)
}
\description{
Outsource the two for loops in novelist_cv into C++
}
