% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novelist_cv.R
\name{novelist_cv}
\alias{novelist_cv}
\title{Rolling Cross-Validation for NOVELIST Threshold Selection}
\usage{
novelist_cv(
  y,
  y_hat,
  S,
  window_size,
  deltas = seq(0, 1, by = 0.05),
  h = 1,
  zero_mean = TRUE,
  error_metric = function(actual, fc) mean((actual - fc)^2, na.rm = TRUE),
  ensure_PD = TRUE,
  message = TRUE
)
}
\arguments{
\item{y}{A matrix of actual values (dim T by p),}

\item{y_hat}{A matrix of fitted values (dim T by p).}

\item{S}{A matrix of reconciliation structure (dim p by b; b is number of bottom series).}

\item{window_size}{The length of the rolling window n.}

\item{deltas}{A numeric vector of candidate threshold values in [0,1].}

\item{h}{An integer vector defining the forecast horizon (default is 1). 
1-to-4-step ahead means h = 1:4; or only 4-step ahead means h = 4.}

\item{zero_mean}{Logical, whether to treat the residuals as zero mean in the covariance.}

\item{error_metric}{An error measure function for given actual and reconciled forecasts.}

\item{ensure_PD}{Logical, whether to ensure the covariance matrix is positive definite.}

\item{reconcile_forecast}{.....}
}
\value{
A list containing:
  \item{delta}{optimal threshold}
  \item{lambda}{optimal shrinkage intensity}
  \item{cov}{optimal NOVELIST covariance matrix}
  \item{errors_by_delta}{vector of average validation error for each delta}
}
\description{
This function performs a rolling window approach to select the threshold 'delta'
for the NOVELIST covariance estimator by minimizing some measure of forecast error
on validation windows.
}
\details{
This approach loops through each possible validation segment in a rolling manner:
  for i in window_size:(T-1), use resid[(i-window_size+1):i, ] as training data
  then produce a reconciled forecast for time i+1 and measure the error.
The final best delta is the one that yields the lowest error measure.
}
