% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{novelist_cv_cpp}
\alias{novelist_cv_cpp}
\title{Rolling Cross-Validation for NOVELIST Threshold Selection (C++ accelerated)}
\usage{
novelist_cv_cpp(
  y,
  y_hat,
  S,
  window_size,
  deltas,
  zero_mean = TRUE,
  ensure_PD = TRUE,
  PD_tol = 1e-06
)
}
\description{
Rolling Cross-Validation for NOVELIST Threshold Selection (C++ accelerated)
}
\note{
This function use different enforcing PD method.
}
\seealso{
[novelist_cv()]
}
