% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{novelist_est_cpp}
\alias{novelist_est_cpp}
\title{NOVELIST covariance estimator, Rcpp version}
\usage{
novelist_est_cpp(resid, delta, lambda_in = NULL, zero_mean = TRUE)
}
\description{
NOVELIST covariance estimator, Rcpp version
}
\seealso{
[novelist_est()]
}
