% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novelist_cv.R
\name{novelist_pc_cv}
\alias{novelist_pc_cv}
\title{Rolling Cross-Validation for NOVELIST with PCA Threshold Selection}
\usage{
novelist_pc_cv(
  y,
  y_hat,
  Ks = 1,
  S,
  window_size,
  deltas = seq(0, 1, by = 0.05),
  h = 1,
  zero_mean = TRUE,
  error_metric = function(actual, fc) mean((actual - fc)^2, na.rm = TRUE),
  ensure_PD = TRUE,
  message = TRUE
)
}
\arguments{
\item{y}{A matrix of actual values (dim T by p),}

\item{y_hat}{A matrix of fitted values (dim T by p).}

\item{Ks}{A numeric vector of candidate number of factors to extract from PCA.
If there is 0, it will not apply PCA and use the full residuals.
Default is 1.}

\item{S}{A matrix of reconciliation structure (dim p by b; b is number of bottom series).}

\item{window_size}{The length of the rolling window n.}

\item{deltas}{A numeric vector of candidate threshold values in [0,1].}

\item{h}{The forecast horizon from 1 to h (default is 1).}

\item{zero_mean}{Logical, whether to treat the residuals as zero mean in the covariance.}

\item{error_metric}{An error measure function for given actual and reconciled forecasts.}

\item{ensure_PD}{Logical, whether to ensure the covariance matrix is positive definite.}

\item{reconcile_forecast}{.....}
}
\value{
A list containing:
 \item{delta}{vector of optimal threshold for each K}
 \item{lambda}{vector of optimal shrinkage intensity for each K}
 \item{cov}{list of optimal NOVELIST covariance matrices for each K}
 \item{errors}{matrix of average validation error for each delta and K}
 \item{ranking_K}{vector of minimum mean errors for each K}
}
\description{
This function performs a rolling window approach to select the threshold 'delta'.
For each window, it applies PCA to the residuals to extract K largest factors,
then applies NOVELIST covariance estimation on the remaindar part.
}
\details{
It can take in multiple candidate K values, and, for each K, returns the best 
threshold 'delta' based on minimising some measure of forecast error on validation windows.


Write soon
}
