% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_estimators.R
\name{novelist_pc_est}
\alias{novelist_pc_est}
\title{NOVELIST Estimator for Covariance Matrix with PCA}
\usage{
novelist_pc_est(
  resid,
  K = 1,
  delta,
  lambda = NULL,
  zero_mean = TRUE,
  ensure_PD = TRUE
)
}
\arguments{
\item{resid}{A numeric matrix of residuals.}

\item{K}{Integer; number of principal components to extract.}

\item{delta}{Numeric; threshold value [0, 1] applied to off-diagonal elements.}

\item{lambda}{Numeric; shrinkage intensity in [0, 1]. If missing, it is computed.}

\item{zero_mean}{Logical; if TRUE, residuals are assumed to have zero mean. Default is TRUE.}

\item{ensure_PD}{Logical; if TRUE, ensures the covariance matrix is positive definite.}
}
\value{
A list containing:
  \item{lambda}{the optimal shrinkage intensity}
  \item{cov}{the NOVELIST covariance matrix}
}
\description{
Estimates a covariance matrix using the NOVELIST approach with PC adjustment.
This function first performs PCA on the residuals to extract K largest PCs,
then applies the NOVELIST estimator on the remaindar part.
}
\seealso{
\code{\link{novelist_est}} for the original NOVELIST estimator.
}
