% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_estimators.R
\name{pc_estimator}
\alias{pc_estimator}
\title{Helper: PC-adjusted Estimators}
\usage{
pc_estimator(resid, estimator_fun, K, standardise, zero_mean)
}
\arguments{
\item{estimator_fun}{A function that takes residuals and returns a covariance matrix.}
}
\value{
A list containing:
  \item{lambda}{the optimal shrinkage intensity from estimator_func}
  \item{cov}{the covariance matrix with PCA adjustment}
}
\description{
A helper function handles PCA decomposition and reconstruction for
PC-adjusted covariance estimators.
}
