% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile.R
\name{reconcile_mint}
\alias{reconcile_mint}
\title{Reconcile Forecasts using Min Trace approach}
\usage{
reconcile_mint(base_forecasts, S, W)
}
\arguments{
\item{base_forecasts}{A matrix of base forecasts (dim T by p).}

\item{S}{A matrix of reconciliation weights (dim p by b, b is number of bottom series).}

\item{W}{A covariance matrix of h-step base forecast errors (dim p by p).}
}
\description{
Reconcile Forecasts using Min Trace approach
}
