% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reconcile_mint_cpp}
\alias{reconcile_mint_cpp}
\title{Reconcile forecasts using MinT formula in C++}
\usage{
reconcile_mint_cpp(base_forecasts, S, W)
}
\description{
Reconcile forecasts using MinT formula in C++
}
\seealso{
\code{\link[=reconcile_mint]}
}
