% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_estimators.R
\name{shrinkage_pc_est}
\alias{shrinkage_pc_est}
\title{Shrinkage Estimator for Covariance Matrix with PCA}
\usage{
shrinkage_pc_est(resid, K = 1, lambda = NULL, zero_mean = TRUE)
}
\arguments{
\item{resid}{A numeric matrix of residuals.}

\item{K}{Integer; number of principal components to extract.}

\item{lambda}{Numeric; shrinkage intensity in [0, 1]. If missing, it is computed.}

\item{zero_mean}{Logical; if TRUE, residuals are assumed to have zero mean. Default is TRUE.}
}
\value{
A list containing:
  \item{lambda}{the optimal shrinkage intensity}
  \item{cov}{the shrinkage covariance matrix with PCA adjustment}
}
\description{
Estimates a covariance matrix using a shrinkage approach with PCA adjustment.
This function first performs PCA on the residuals to extract K largest PCs,
then applies the shrinkage estimator on the remainder part.
}
\seealso{
\code{\link{shrinkage_est}} for the original shrinkage estimator.
}
