% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_sim_result.R
\name{transform_error_list}
\alias{transform_error_list}
\title{Transform an length-M list of models' error2 matrices into df format}
\usage{
transform_error_list(error_list)
}
\arguments{
\item{error_list}{A length-M list of list of error matrices where each matrix
corresponds to a model's error squared values.}
}
\value{
A data frame containing the error squared values, model names,
        series names, and forecast horizons, with id represent simulation rep.
}
\description{
Transform an length-M list of models' error2 matrices into df format
}
\examples{
error_list <- list(
  base = matrix(c(0.1, 0.2, 0.3, 0.4, 0.5, 0.6), nrow = 3),
  mint_shr = matrix(c(0.2, 0.3, 0.4, 0.5, 0.6, 0.7), nrow = 3)
)
error_df <- transform_error_list(error_list)

}
