% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_sim_result.R
\name{transfrom_sim_MSE}
\alias{transfrom_sim_MSE}
\title{Transform a list of MSE matrices into a tsibble/tibble format}
\usage{
transfrom_sim_MSE(MSE, to_ts = T)
}
\arguments{
\item{MSE}{A list of matrices where each matrix corresponds to a model's MSE values.}

\item{to_ts}{Logical indicating whether to convert the result to a tsibble.}
}
\value{
A tsibble or tibble containing the MSE values, model names, series names, and forecast horizons.
}
\description{
Transform a list of MSE matrices into a tsibble/tibble format
}
\examples{
MSE <- list(
  model1 = matrix(c(0.1, 0.2, 0.3, 0.4, 0.5, 0.6), nrow = 3),
  model2 = matrix(c(0.2, 0.3, 0.4, 0.5, 0.6, 0.7), nrow = 3)
)
result <- transfrom_sim_MSE(MSE, to_ts = TRUE)

}
