% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{convert_cor2cov}
\alias{convert_cor2cov}
\title{Convert correlation to covariance matrix}
\usage{
convert_cor2cov(cor, stdevs = runif(nrow(cor), sqrt(2), sqrt(6)))
}
\description{
Convert correlation to covariance matrix
}
