% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot a heatmap of a matrix}
\usage{
plot_heatmap(mat, is_unit_spherical = FALSE)
}
\arguments{
\item{mat}{A matrix to be plotted as a heatmap.}

\item{is_unit_spherical}{Logical, if TRUE, the color scale is set to [-1, 1].}
}
\value{
A ggplot object representing the heatmap.
}
\description{
Plot a heatmap of a matrix
}
\examples{
mat <- matrix(rnorm(100), nrow = 10)
plot_heatmap(mat)
plot_heatmap(mat, is_unit_spherical = TRUE)

}
